app/requirements.txt
fastapi
uvicorn[standard]
onnxruntime
sentence-transformers